# packages/shared/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "Buscador y Visor de Documentos Tecnicos",
    "description" : "Carga, indexacion asincrona, busqueda full-text y visualizacion de documentos",
    "version" : "0.1.0"
  },
  "paths" : {
...
    "/search" : {
      "get" : {
        "tags" : [ "search-controller" ],
        "summary" : "Search indexed documents",
        "description" : "Quoted phrases, `or` between alternatives and a leading minus to exclude a word are all understood. Only documents that finished indexing are ever returned.",
        "operationId" : "search",
        "parameters" : [ {
          "name" : "q",
          "in" : "query",
          "description" : "Words or quoted phrases to look for",
          "required" : false,
          "schema" : {
            "type" : "string",
            "default" : ""
          },
          "example" : "indice invertido -borrador"
        }, {
          "name" : "page",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 1
          }
        }, {
          "name" : "pageSize",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 10
          }
        }, {
          "name" : "category",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "enum" : [ "MANUAL", "SPECIFICATION", "ARCHITECTURE_GUIDE", "OTHER" ]
          }
        }, {
          "name" : "author",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "tags",
          "in" : "query",
          "description" : "Every tag given must be present on the document",
          "required" : false,
          "schema" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "A page of ranked matches with highlighted fragments",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SearchResponse"
                }
              }
            }
          },
          "400" : {
            "description" : "The query carries no searchable term, or the paging is out of range",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "503" : {
            "description" : "The search exceeded its time budget and was cancelled",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    },
...
    "/documents/{id}/content" : {
      "get" : {
        "tags" : [ "document-catalog-controller" ],
        "summary" : "Read a document's body",
        "description" : "Chunks in order, paged by a cursor over chunkIndex.",
        "operationId" : "content",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "fromChunkIndex",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 0
          }
        }, {
          "name" : "limit",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 50
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/DocumentContentResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components" : {
    "schemas" : {
      "ProblemDetail" : {
        "type" : "object",
        "description" : "RFC 9457 problem detail",
        "properties" : {
          "type" : {
            "type" : "string",
            "format" : "uri",
            "example" : "urn:problem-type:invalid-upload"
          },
          "title" : {
            "type" : "string",
            "example" : "Unprocessable Entity"
          },
          "status" : {
            "type" : "integer",
            "format" : "int32",
            "example" : 422
          },
          "detail" : {
            "type" : "string",
            "description" : "What went wrong, in English; the interface shows its own wording"
          },
          "instance" : {
            "type" : "string",
            "format" : "uri",
            "example" : "/api/documents"
          },
          "requestId" : {
            "type" : "string",
            "description" : "Correlates the failure with the server logs"
          },
          "errors" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/UploadFileProblem"
            }
          }
        },
        "required" : [ "status", "title", "type" ]
      },
      "UploadFileProblem" : {
        "type" : "object",
        "description" : "Why one file of a batch was turned down",
        "properties" : {
          "index" : {
            "type" : "integer",
            "format" : "int32",
            "example" : 1
          },
          "filename" : {
            "type" : "string",
            "example" : "manual.exe"
          },
          "rule" : {
            "type" : "string",
            "example" : "EXTENSION_ALLOWLIST"
          },
          "errorCode" : {
            "type" : "string",
            "example" : "UNSUPPORTED_FORMAT"
          },
          "existingDocumentId" : {
            "type" : "string",
            "description" : "The document that already holds the same content"
          }
        },
        "required" : [ "index" ]
      },
...
      "SearchHitResponse" : {
        "type" : "object",
        "description" : "One matching document with its best fragment",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "title" : {
            "type" : "string"
          },
          "titleHighlight" : {
            "type" : "string",
            "description" : "The title with every matched word wrapped in the U+27E6 and U+27E7 sentinels",
            "example" : "Guia de ⟦despliegue⟧"
          },
          "author" : {
            "type" : "string"
          },
          "category" : {
            "type" : "string",
            "enum" : [ "MANUAL", "SPECIFICATION", "ARCHITECTURE_GUIDE", "OTHER" ]
          },
          "tags" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "version" : {
            "type" : "string"
          },
          "indexedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "chunkIndex" : {
            "type" : "integer",
            "format" : "int32",
            "description" : "Index of the chunk the fragment came from, so the viewer can scroll to it"
          },
          "page" : {
            "type" : "integer",
            "format" : "int32",
            "description" : "Page the fragment came from, for documents that have pages"
          },
          "heading" : {
            "type" : "string"
          },
          "snippet" : {
            "type" : "string",
            "description" : "Matched fragment, with the same sentinels as the title"
          },
          "rank" : {
            "type" : "number",
            "format" : "double"
          }
        }
      },
      "SearchResponse" : {
        "type" : "object",
        "description" : "One page of search results, highest ranked first",
        "properties" : {
          "items" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/SearchHitResponse"
            }
          },
          "total" : {
            "type" : "integer",
            "format" : "int64",
            "description" : "Documents matching the query and the filters, not chunks"
          },
          "page" : {
            "type" : "integer",
            "format" : "int32"
          },
          "pageSize" : {
            "type" : "integer",
            "format" : "int32"
          },
          "tookMs" : {
            "type" : "integer",
            "format" : "int64",
            "description" : "How long the engine took, excluding rendering this response"
          }
        }
      },
...
      "DocumentChunkResponse" : {
        "type" : "object",
        "description" : "One slice of a document's body",
        "properties" : {
          "chunkIndex" : {
            "type" : "integer",
            "format" : "int32"
          },
          "page" : {
            "type" : "integer",
            "format" : "int32"
          },
          "heading" : {
            "type" : "string"
          },
          "content" : {
            "type" : "string"
          }
        }
      },
      "DocumentContentResponse" : {
        "type" : "object",
        "description" : "A window over a document's chunks, paged by a cursor over chunkIndex",
        "properties" : {
          "chunks" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/DocumentChunkResponse"
            }
          },
          "nextChunkIndex" : {
            "type" : "integer",
            "format" : "int32",
            "description" : "Cursor for the next window, absent when the document ends here"
          }
        }
      }
    }
  }
}